requirements.txt
Django
djangorestframework
django-crispy-forms
Pillow